Interceptor.init is called at runtime, and return a control-flow tuple
When setting up the adaptor, or otherwise using an Endpoint,
`Endpoint.interceptors/1` is called which readies both endpoint and
server interceptors.

These are then iterated over per-request, and the decision of whether to
continue on or not is made based on the return value from the
interceptor's `call` function.

Interceptors may return a :cont, :halt, :after, or :error tagged tuple
which determines when the next interceptor is run or not.

The primary benefit of this change is that `init/1` calls are always
made at runtime, and not in a macro, which loosens a lot of
restrictions.

This makes it easy, for instance, to add the allowed methods header,
with configuration, to the CORS interceptor.